pyproject.toml
[project]
name = "pantherhacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
]
